+++ fixtures/low-level/src/index.ts
@@ -1,8 +1,10 @@
-import { block } from './styles.css';
+import { block, container } from './styles.css';
 import testNodes from '../test-nodes.json';
 
 document.body.innerHTML = `
+<div class="${container}"> 
   <div id="${testNodes.block}" class="${block}"> 
     I'm a block
   </div>
+</div>
 `;