+++ packages/css/src/entries/fileScope.ts
@@ -31,7 +31,7 @@ export function getFileScope(): FileScope {
         Styles were unable to be assigned to a file. This is generally caused by one of the following:
 
         - You may have created styles outside of a '.css.ts' context
-        - You may have incorrect configuration. See https://vanilla-extract.style/documentation/setup
+        - You may have incorrect configuration. See https://vanilla-extract.style/documentation/getting-started
       `,
     );
   }