+++ packages/css/src/identifier.ts
@@ -3,23 +3,46 @@ import hash from '@emotion/hash';
 import { getIdentOption } from './entries/adapter';
 import { getAndIncrementRefCounter, getFileScope } from './entries/fileScope';
 
-function getDevPrefix(debugId: string | undefined) {
+function getDevPrefix({
+  debugId,
+  debugFileName,
+}: {
+  debugId?: string;
+  debugFileName: boolean;
+}) {
   const parts = debugId ? [debugId.replace(/\s/g, '_')] : [];
-  const { filePath } = getFileScope();
 
-  const matches = filePath.match(
-    /(?<dir>[^\/\\]*)?[\/\\]?(?<file>[^\/\\]*)\.css\.(ts|js|tsx|jsx)$/,
-  );
+  if (debugFileName) {
+    const { filePath } = getFileScope();
 
-  if (matches && matches.groups) {
-    const { dir, file } = matches.groups;
-    parts.unshift(file && file !== 'index' ? file : dir);
+    const matches = filePath.match(
+      /(?<dir>[^\/\\]*)?[\/\\]?(?<file>[^\/\\]*)\.css\.(ts|js|tsx|jsx|cjs|mjs)$/,
+    );
+
+    if (matches && matches.groups) {
+      const { dir, file } = matches.groups;
+      parts.unshift(file && file !== 'index' ? file : dir);
+    }
   }
 
   return parts.join('_');
 }
 
-export function generateIdentifier(debugId: string | undefined) {
+interface GenerateIdentifierOptions {
+  debugId?: string;
+  debugFileName?: boolean;
+}
+
+export function generateIdentifier(debugId?: string): string;
+export function generateIdentifier(options?: GenerateIdentifierOptions): string;
+export function generateIdentifier(
+  arg?: string | GenerateIdentifierOptions,
+): string {
+  const { debugId, debugFileName = true } = {
+    ...(typeof arg === 'string' ? { debugId: arg } : null),
+    ...(typeof arg === 'object' ? arg : null),
+  };
+
   // Convert ref count to base 36 for optimal hash lengths
   const refCount = getAndIncrementRefCounter().toString(36);
   const { filePath, packageName } = getFileScope();
@@ -31,7 +54,7 @@ export function generateIdentifier(debugId: string | undefined) {
   let identifier = `${fileScopeHash}${refCount}`;
 
   if (getIdentOption() === 'debug') {
-    const devPrefix = getDevPrefix(debugId);
+    const devPrefix = getDevPrefix({ debugId, debugFileName });
 
     if (devPrefix) {
       identifier = `${devPrefix}__${identifier}`;