+++ packages/css/src/transformCss.test.ts
@@ -961,7 +961,41 @@ describe('transformCss', () => {
     `);
   });
 
-  it('should handle nested @supports and @media queries', () => {
+  it('should handle @container queries', () => {
+    expect(
+      transformCss({
+        composedClassLists: [],
+        localClassNames: ['testClass'],
+        cssObjs: [
+          {
+            type: 'local',
+            selector: 'testClass',
+            rule: {
+              display: 'flex',
+              containerName: 'sidebar',
+              '@container': {
+                'sidebar (min-width: 700px)': {
+                  display: 'grid',
+                },
+              },
+            },
+          },
+        ],
+      }).join('\n'),
+    ).toMatchInlineSnapshot(`
+      ".testClass {
+        display: flex;
+        container-name: sidebar;
+      }
+      @container sidebar (min-width: 700px) {
+        .testClass {
+          display: grid;
+        }
+      }"
+    `);
+  });
+
+  it('should handle nested @supports, @media and @container queries', () => {
     expect(
       transformCss({
         composedClassLists: [],
@@ -978,16 +1012,27 @@ describe('transformCss', () => {
                   '@media': {
                     'screen and (min-width: 700px)': {
                       display: 'grid',
+                      '@container': {
+                        'sidebar (min-width: 700px)': {
+                          display: 'grid',
+                        },
+                      },
                     },
                   },
                 },
               },
+
               '@media': {
                 'screen and (min-width: 700px)': {
                   color: 'green',
                   '@supports': {
                     '(display: grid)': {
                       borderColor: 'blue',
+                      '@container': {
+                        'sidebar (min-width: 700px)': {
+                          display: 'grid',
+                        },
+                      },
                     },
                   },
                 },
@@ -1008,6 +1053,11 @@ describe('transformCss', () => {
           .testClass {
             border-color: blue;
           }
+          @container sidebar (min-width: 700px) {
+            .testClass {
+              display: grid;
+            }
+          }
         }
       }
       @supports (display: grid) {
@@ -1018,12 +1068,17 @@ describe('transformCss', () => {
           .testClass {
             display: grid;
           }
+          @container sidebar (min-width: 700px) {
+            .testClass {
+              display: grid;
+            }
+          }
         }
       }"
     `);
   });
 
-  it('should merge nested @supports and @media queries', () => {
+  it('should merge nested @supports, @media and @container queries', () => {
     expect(
       transformCss({
         composedClassLists: [],
@@ -1038,12 +1093,18 @@ describe('transformCss', () => {
                   '@supports': {
                     '(display: grid)': {
                       borderColor: 'blue',
+                      '@container': {
+                        'sidebar (min-width: 700px)': {
+                          display: 'grid',
+                        },
+                      },
                     },
                   },
                 },
               },
             },
           },
+
           {
             type: 'local',
             selector: 'otherClass',
@@ -1053,6 +1114,11 @@ describe('transformCss', () => {
                   '@supports': {
                     '(display: grid)': {
                       backgroundColor: 'yellow',
+                      '@container': {
+                        'sidebar (min-width: 700px)': {
+                          display: 'grid',
+                        },
+                      },
                     },
                   },
                 },
@@ -1069,6 +1135,14 @@ describe('transformCss', () => {
           }
           .otherClass {
             background-color: yellow;
+          }
+          @container sidebar (min-width: 700px) {
+            .testClass {
+              display: grid;
+            }
+            .otherClass {
+              display: grid;
+            }
           }
         }
       }"