+++ playwright.config.ts
@@ -5,7 +5,8 @@ const config: PlaywrightTestConfig = {
   updateSnapshots: 'none',
   expect: {
     toMatchSnapshot: {
-      threshold: 0.2,
+      threshold: 0.3,
+      maxDiffPixelRatio: 0.05,
     },
   },